Add files via upload
This micro project embarks on a compelling journey into the realms of Facebook 
data, wielding the analytical prowess of the R programming language. R, 
renowned for its statistical rigor and versatility in data manipulation, emerges as 
the ideal tool to unravel the intricate patterns and nuances concealed within the 
vast troves of social media data. Through this exploration, we aim to extract 
meaningful insights, providing a nuanced understanding of user behaviors, 
content dynamics, and the evolving landscape of online social connections.